canvas and table rendering are object methods
Pass a reference to a canvas dom object when creating a new model:

  avalanche2d.Model(canvas, model_options, array_selection)

This can be null if you are not going to use the rendering.

Re-organize source so the utilities functions are both
at the end and are private to the avalanche2d module.

setupRGBAColorTables is now private and initialized when
a new model is created.